+++ src/module/data/message/_module.mjs
@@ -1,3 +1,2 @@
 export { default as BaseMessageModel } from "./base.mjs";
-export { default as ProjectRollModel } from "./project-roll.mjs";
 export { default as StandardModel } from "./standard.mjs";