+++ src/module/data/pseudo-documents/message-parts/_module.mjs
@@ -6,3 +6,4 @@ export { default as HeroTokenPart } from "./hero-token.mjs";
 export { default as RollPart } from "./roll.mjs";
 export { default as SavingThrowPart } from "./saving-throw.mjs";
 export { default as TestPart } from "./test.mjs";
+export { default as ProjectPart } from "./project.mjs";