+++ src/module/data/pseudo-documents/message-parts/_types.d.ts
@@ -34,3 +34,10 @@ declare module "./test.mjs" {
     resultSource: string;
   }
 }
+
+declare module "./project.mjs" {
+  export default interface ProjectPart {
+    projectUuid: string;
+    event: boolean;
+  }
+}